set_channel_by_refnum() is the means by which we set the current_channel for a window, but due to my malfeasance, several bugs slipped through.
Because referential integrity panic failures will occur if you incorrectly
set a channel to a window, and because it's very hard to track down the
cause of that after it's already happened, make set_channel_by_refnum()
sanity check it's arguments and panic immediately if asked to do something
invalid.

Specifically, a window cannot have as its current channel a channel that
has not yet been joined (it is a "waiting_channel" until we get the JOIN
back), so set_channel_by_refnum() must refuse to make a channel the current
channel if it does not, in fact, exist.

Also, we must not default to the current window if the specified refnum
is invalid.  There is no reasonable circumstance in which an invalid
window refnum could be passed to set_channel_by_refnum().  So if the window
refnum lookup fails, abort immediately.

These two changes should make it basically impossible to sneak a current
channel onto an invalid window (ie, a window on another server), or to
"accidentally" use a server refnum where a window refnum was needed.